Renamed indexer file.
