+++ src/indexer/indexer-functions.ts
